init: use correct names for kubernetes resource fields
Previously, the `init` functionality was just using the passed-in stack
name (which is expected to be in dockerfile image name format, like
organization/image), but kubernetes resource fields don't like forward
slashes in the values. This changeset converts all forward slashes to
hyphens before using them in kubernetes resource fields.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>