init: add support for spinning up the docker registry, fix broken reference
This changeset fixes a broken reference which was from the kubebuilder
boilerplate; it was preventing the full build+publish cycle from
working. It also adds docker-push for publishing. It *also* adds support
for spinning up a local docker registry (if one isn't already running)
for use during development and local testing of a stack.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>