init: pull rbac from kubebuilder role instead of separate file
Previously, the stack build was pulling the rbac.yaml from an
independently-managed `rbac.yaml`. This file contained essentially the
same information as the kubebuilder project's `role.yaml`, but using
just the `rules` field. It turns out the stack manager will accept a
full role definition, so managing a separate but slightly different
`rbac.yaml` is no longer necessary; the kubebuilder-generated
`role.yaml` can be used directly. The new logic pulls the `role.yaml`,
but will overwrite it with the separate `rbac.yaml` if it exists.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>