uninstall: use `shift` after consuming the stack name argument
The `uninstall` command passes arguments through to `kubectl`, and if we
don't `shift` the stack name out, it'll be passed twice into the
`kubectl` command. The reason we're separating them at all is to make
the code more readable, and because the passthrough behavior may be
tweaked in the future.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>