cli: split install yaml generation into lower-level command
Previously, the logic to generate the yaml to pass to kubectl on install
was inside the same command as the rest of the install logic. Now, it's
in its own lower-level command, which means it can be reused for
different purposes other than just straight installing. For example, if
someone wants to generate the yaml and edit it before using it, it's
much more convenient if the lower-level command exists.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>