commands: add list command
When interacting with the wordpress stack, it seemed useful to be able
to list all installed stacks, so this changeset adds a `list` command
which does that. The default behavior is to list stacks in all
namespaces. If any arguments are passed, it'll use those instead (and
pass them through to `kubectl`)

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>