stack init: add --force option
It is possible to run kubectl crossplane stack init from the wrong
path or as a repeated command and wind up losing all of the existing
stack information.

This patch adds a "stack init -f" (--force) argument that must be
supplied to allow an existing stack directory to be re-init'd.

The "cd" command issued during "init" to enter the install _directory_
is now prevented from interpretting _directory_ as arguments for "cd".

Also included is a minor change to "make install" which permits
"make install" to be re-run (useful during local development iterations).

Signed-off-by: Marques Johansson <marques@upbound.io>