trace: fix finding related provider of a resource class
Full object reference from resource class to provider is not available.
Only name of the provider there. To resolve the provider, there was a
workaround in the code, which sets apiVersion of the provider as the
apiVersion of the resource class. This assumption does not hold with
v1beta1 versions recently introduced in stack-gcp. This commit fixes
by setting group instead of apiVersion.

Signed-off-by: Hasan Turken <turkenh@gmail.com>